#50 refactor: scope-event-context を _events/_contexts へ移動
- Context 定義を time-control-03 直下ではなく _events 内部の _contexts/ に配置し、_events モジュールの自己完結性を保つ
