+++ src/prototypes/time-control/time-control-03/_events/_hooks/_utils/use-time-control-03-event-dispatcher.ts
@@ -1,7 +1,7 @@
 import { useEventDispatcher } from '@/hooks/use-event-dispatcher'
 
+import { useScopeEventTarget } from '../../_contexts/scope-event-context'
 import { TimeControl03EventMap } from '../../index.types'
-import { useScopeEventTarget } from '../../scope-event-context'
 
 /**
  * scope (この time-control-03 instance) 専用 EventTarget へイベントを発行する dispatcher を返す
